+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Fix bug causing incorrect nvidia-smi symlink to be created on WSL2 systems with multiple driver roots.
 * Fix bug when using driver versions that do not include a patch component in their version number.
 * Skip additional modifications in CDI mode.
+* Fix loading of kernel modules and creation of device nodes in containerized use cases.
 
 * [toolkit-container] Allow same envars for all runtime configs